+++ include/Firefly/Core/Application.hh
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <string>
+#include <memory>
 
 #include "Window.hh"
 #include "LayerStack.hh"
