+++ include/Firefly/Core/Events/Event.hh
@@ -3,7 +3,7 @@
 #include <string>
 #include <functional>
 
-#include "../Base.hh"
+#include "../../Base.hh"
 
 namespace flyCore {
