+++ include/Firefly/Core/Events/KeyEvent.hh
@@ -40,7 +40,7 @@ namespace flyCore {
         std::string ToString() const override
         {
             std::stringstream ss;
-            ss << "KeyPressedEvent: " << m_KeyCode << " (" << m_RepeatCount << "repeats)";
+            ss << "KeyPressedEvent: " << m_KeyCode << " (" << m_RepeatCount << " repeats)";
             return ss.str();
         }
