+++ include/Firefly/Firefly.hh
@@ -2,11 +2,9 @@
 
 // For use by firefly applications
 
-// #include "Core/Application.hh"
-// #include "Core/Log.hh"
-#include "../src/Firefly-Core/Application.hh"
-#include "../src/Firefly-Core/Log.hh"
-#include "../src/Firefly-Core/Layer.hh"
+#include "Core/Application.hh"
+#include "Core/Log.hh"
+#include "Core/Layer.hh"
 
 // ----Entry Point----------------------------
 #include "EntryPoint.hh"
