+++ include/Firefly/PlatformDetection.hh
