+++ src/Firefly-Core/Application.cc
@@ -2,6 +2,8 @@
 #include "Log.hh"
 #include "Events/ApplicationEvent.hh"
 
+#include "glad/glad.h"
+
 namespace flyCore {
 
     Application::Application()
@@ -49,6 +51,9 @@ namespace flyCore {
     {
         while (m_Running)
         {
+            glClearColor(1, 0, 1, 1);
+            glClear(GL_COLOR_BUFFER_BIT);
+
             for (Layer* layer : m_LayerStack)
                 layer->OnUpdate();
