+++ src/Firefly-Core/LayerStack.cc
@@ -1,4 +1,4 @@
-#include "LayerStack.hh"
+#include "Firefly/Core/LayerStack.hh"
 #include <algorithm>
 
 namespace flyCore {
